Move `next_hour` helper to common test helpers file
